Refactor base.py to be a little easier to follow.
Summary:
My brain was exploding with all the terminology (cache, local_cache, _cache, etc), so I refactored the variable names and whatnot to be a little more semantic.

NOTE: this does **not** fix the cache hit performance regression, but this was a dependency on it to make the logic actually read and make sense.

Test Plan: existing

Reviewers: dcramer

Reviewed By: dcramer

Differential Revision: http://phabricator.local.disqus.net/D4132